Added tidetable methods and basic tutorial
